Add flags for shader, transition, effect and bitmap generator data to tell whether the old JSON can be deleted or not
Right now, since the data can only ever come from browser_canvas, which generates a new cJSON* object for every request, this is always the case and is consistent with previous behavior.
However, in the future, data might come from other sources (e.g. a file in SPIFFS) which must not be deleted after use.